+++ benchmarks/internal/benchmark-utils.js
@@ -25,9 +25,25 @@ function generateRandomInteger (max, min = 0) {
   return Math.floor(Math.random() * (max + 1))
 }
 
+/**
+ * Intentionally inefficient implementation.
+ *
+ * @param {*} n
+ * @returns
+ */
+function fibonacci (n) {
+  if (n <= 1) return 1
+  return fibonacci(n - 1) + fibonacci(n - 2)
+}
+
 const LIST_FORMATTER = new Intl.ListFormat('en-US', {
   style: 'long',
   type: 'conjunction'
 })
 
-module.exports = { generateRandomInteger, LIST_FORMATTER, runPoolifierTest }
+module.exports = {
+  generateRandomInteger,
+  LIST_FORMATTER,
+  fibonacci,
+  runPoolifierTest
+}