+++ src/index.ts
@@ -1,16 +1,15 @@
-export type {
-  ErrorHandler,
-  ExitHandler,
-  IWorker,
-  OnlineHandler,
-  PoolOptions
-} from './pools/abstract-pool'
 export { DynamicClusterPool } from './pools/cluster/dynamic'
 export { FixedClusterPool } from './pools/cluster/fixed'
 export type { ClusterPoolOptions } from './pools/cluster/fixed'
-export type { IPool } from './pools/pool'
-export { WorkerChoiceStrategies } from './pools/selection-strategies'
-export type { WorkerChoiceStrategy } from './pools/selection-strategies'
+export type { IPool, PoolOptions } from './pools/pool'
+export type {
+  ErrorHandler,
+  ExitHandler,
+  IPoolWorker,
+  OnlineHandler
+} from './pools/pool-worker'
+export { WorkerChoiceStrategies } from './pools/selection-strategies/selection-strategies-types'
+export type { WorkerChoiceStrategy } from './pools/selection-strategies/selection-strategies-types'
 export { DynamicThreadPool } from './pools/thread/dynamic'
 export { FixedThreadPool } from './pools/thread/fixed'
 export type { ThreadWorkerWithMessageChannel } from './pools/thread/fixed'